server/pyproject.toml
[tool.poetry]
name = "scally"
version = "0.1.0"
description = ""
authors = ["Shrey Pandya <shrey150@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
langchain-community = "^0.2.1"
python-dotenv = "^1.0.1"
gitpython = "^3.1.43"
langchain-chroma = "^0.1.1"
langchain-openai = "^0.1.8"
sentence-transformers = "^3.0.0"
fastapi = "^0.111.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
